+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 * Add scope to all EditCustomFieldsSettings and ViewCustomFieldsSettings. Refs UIOR-1516.
 * Rename permissions to make it easier to differentiate between them. Refs UIOR-1504.
 * Fix permission error for edit custom-fields. Refs UIOR-1513.
+* Reset selected FY when PO has changed. Refs UIOR-1510.
 
 ## [8.0.5](https://github.com/folio-org/ui-orders/tree/v8.0.5) (2025-06-30)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v8.0.4...v8.0.5)